Removed "MEGA" challenges types, moved ChallengeType in Challenge class and fixed spigot-only ChatColor import
